BufferOutputStream: do not return the only encoding out of spite
My diagnosis of commit d431111 was wrong. If all encoding detections
failed, we were returning the first of the list (even knowing it wrong),
not the last one.
And if the list contained only 1 encoding, we were returning it without
even testing it.
Fix this and make so that the fallback detection properly check every
encoding of the list and returns NULL appropriately when no charset
matches.